Fix snapshot/delta clean-ups.

+++ src/main/scala/net/ripe/rpki/publicationserver/store/fs/RemovingFileVisitor.scala
@@ -1,12 +1,14 @@
 package net.ripe.rpki.publicationserver.store.fs
 
 import java.io.IOException
-import java.nio.file.{Files, FileVisitResult, Path, SimpleFileVisitor}
+import java.nio.file._
 import java.nio.file.attribute.{BasicFileAttributes, FileTime}
 
 import net.ripe.rpki.publicationserver.Logging
 
-class RemovingFileVisitor(timestamp: FileTime, filenameToDelete: Path) extends SimpleFileVisitor[Path] with Logging {
+class RemovingFileVisitor(timestamp: FileTime, filenameToDelete: Path, latestSerial: Long) extends SimpleFileVisitor[Path] with Logging {
+
+  val latestSnapshotNameComponent = Paths.get(latestSerial.toString, filenameToDelete.toString)
 
   override def visitFileFailed(file: Path, exc: IOException): FileVisitResult = {
     logger.error(s"Error visiting $file: ${exc.getMessage}")
@@ -15,7 +17,7 @@ class RemovingFileVisitor(timestamp: FileTime, filenameToDelete: Path) extends S
   }
 
   override def visitFile(file: Path, attrs: BasicFileAttributes): FileVisitResult = {
-    if (file.endsWith(filenameToDelete) && isModifiedBefore(file)) {
+    if (file.endsWith(filenameToDelete) && !file.endsWith(latestSnapshotNameComponent) && isModifiedBefore(file)) {
       logger.info(s"Removing $file")
       Files.deleteIfExists(file)
     }
@@ -31,11 +33,11 @@ class RemovingFileVisitor(timestamp: FileTime, filenameToDelete: Path) extends S
   }
 
   def isModifiedBefore(file: Path): Boolean = {
-    Files.getLastModifiedTime(file).compareTo(timestamp) < 0
+    Files.getLastModifiedTime(file).compareTo(timestamp) <= 0
   }
 
   def isCreatedBefore(path: Path): Boolean = {
-    Files.readAttributes(path, classOf[BasicFileAttributes]).creationTime.compareTo(timestamp) < 0
+    Files.readAttributes(path, classOf[BasicFileAttributes]).creationTime.compareTo(timestamp) <= 0
   }
 
   def isEmptyDir(dir: Path): Boolean = {
